requirements.txt
#  Core API
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1

#  Web Interface
flask==3.0.3
requests==2.32.3

# Embeddings / ML 
sentence-transformers==3.0.1
torch==2.3.0

# Vector Database
chromadb==0.5.0
pysqlite3-binary==0.5.3; platform_system != "Windows"

# Clustering / ML Utilities
scikit-learn==1.4.2
numpy==1.26.4
scipy==1.13.0
scikit-fuzzy==0.4.2

#Data Processing 
pandas==2.2.2
tqdm==4.66.4

#Visualization / Notebook
jupyter==1.0.0
matplotlib==3.8.4
seaborn==0.13.2
umap-learn==0.5.6

# Testing
pytest==8.2.2